requirements.txt
click==8.1.7
matplotlib==3.9.2
scikit_learn==1.5.1
torch==2.4.0
torchvision
hydra-core

loguru==0.7.3
wandb==0.19.1
